update the ranking system

+++ src/main/java/com/freeloader/triviaservice/model/HighScoreDetail.java
@@ -1,5 +1,5 @@
 package com.freeloader.triviaservice.model;
 
-public record HighScoreDetail(String name, Long score, String date, int ranking) {
+public record HighScoreDetail(String name, int score, String date, int ranking) {
 
 }
